scripts/requirements.txt
# Requirements for citation update script
requests>=2.28.0
pyyaml>=6.0
scholarly>=1.7.0
